+++ operations/helm/charts/grafana-agent/CHANGELOG.md
@@ -10,6 +10,17 @@ internal API changes are not present.
 Unreleased
 ----------
 
+### Features
+
+- Allow setting nodePort for service. (@ryayon)
+
+0.36.0 (2024-02-27)
+-------------------
+
+### Enhancements
+
+- Update Grafana Agent version to v0.40.2. (@rfratto)
+
 0.35.0 (2024-02-27)
 -------------------